feat: support declarative schema with db diff

+++ internal/db/reset/reset_test.go
@@ -28,7 +28,7 @@ func TestResetCommand(t *testing.T) {
 
 	var dbConfig = pgconn.Config{
 		Host:     utils.Config.Hostname,
-		Port:     uint16(utils.Config.Db.Port),
+		Port:     utils.Config.Db.Port,
 		User:     "admin",
 		Password: "password",
 		Database: "postgres",
